Update build.sh to build libinfra.a before MOM6
If we turn the modules in config_src/infra/FMS2 into a library, then that
library can be linked when building the MOM6 executable but it also means we
can link that library when building unit tests (without building the rest of
MOM6).

Running "build.sh --unit-tests-only" will build libinfra.a without the MOM6
executable, but does not yet contain any unit tests.