Update build.sh to build libinfra.a before MOM6 (#47)
* Update build.sh to build libinfra.a before MOM6

If we turn the modules in config_src/infra/FMS2 into a library, then that
library can be linked when building the MOM6 executable but it also means we
can link that library when building unit tests (without building the rest of
MOM6).

Running "build.sh --unit-tests-only" will build libinfra.a without the MOM6
executable, but does not yet contain any unit tests.

* Order matters when linking

need "-linfra -lFMS" rather than "-lFMS -linfra"; I guess the macOS linker is
more lenient than the linux one

* User can specify which infra dir to use

Default is FMS2, but when we have an AMReX option we can run

./build.sh --infra AMReX

* Updates from code review

More readable / editable list of files for the MOM6 files that are dependencies
of the config_src/infra/FMS2 directory

* More generality in build

Only build FMS if INFRA==FMS2; generalize LINKING_FLAGS so when AMReX
infrastructure is ready we don't try to link to FMS

* No more tickmarks in build.sh

* Abort for invalid infrastructure choice

Currently requires INFRA=FMS2; when we add support for AMReX we'll need to
update the error message in the "else" block